[VCDA-755] Ensure templates catalog is always shared (#108)
Before this fix, if specified catalog already existed, it wouldn't
be shared with other orgs. If specified catalog did not exist,
it would be created, but would not be shared. CSE installation
would crash because we did not reload the org representation after
the POST call.

This fix creates the catalog if needed, always shares the catalog
to other orgs, and properly reloads the org representation

Tested with all CSE commands

Signed-off-by: Andrew Ni <niandrew7@gmail.com>