Create AMQP exchange even if settings are the same
Commit 6fca204be2f257eaf80534712a91850a27453e5e caused a
regression where if the vCD and config AMQP settings were the same,
CSE installation would skip AMQP configuration. However, just
because the settings were the same, doesn't mean the AMQP exchange
exists. The workaround for this was to change the name of the vCD
AMQP exchange to trigger the AMQP configuration option during
CSE install.

Now, if settings are the same, AMQP exchange will still be
checked/created unless the user says 'no' during the prompt

Signed-off-by: Andrew Ni <niandrew7@gmail.com>