Generalize cse registration
register_extension() is now a general function in utils.py, and
can be used to register any extension.

should_register_cse() is a boolean function that decides
whether or not to register cse based on user input. Generalized
this functionality for test convenience, and moved to avoid
polluting config.py

Signed-off-by: Andrew Ni <niandrew7@gmail.com>