Added org and vdc as optional parameters
To accomodate performance needs as well as reduce parameter
redundancy, functions that use org or vdc now take
them as optional parameters. If org/vdc is passed in, the
function will use those orgs, but these functions are adaptable,
so if org/vdc is not passed, it can figure out which org/vdc
it is supposed to use. The priority goes from: given org/vdc
object > client + specific org name > client's logged in org.

Signed-off-by: Andrew Ni <niandrew7@gmail.com>