[VCDA-865] Remove vCD AMQP configuration and Enhance Extension Checks (#206)
We were configuring vCD AMQP settings during CSE installation when we should not. If other extensions are being used, this re-configuring could cause errors in those extensions. This functionality has been removed. (Related to #187)

Fix a bug where if CSE is already registered as an extension to vCD, but has different properties than the config file, then we do not update vCD-CSE properties. Now when installing CSE, if CSE already exists on vCD but has different properties than config file, the user is alerted and presented with an option to update CSE properties on vCD. (Dependent on vmware/pyvcloud#379)

Remove system testing functionality of checking vCD AMQP configuration

Update docs to reflect that we do not configure vCD AMQP anymore

Fixes #187