Add a Dockerfile
Adds an implemenation of a Dockerfile, to deliver the CSE
command line as a single container.

Following VMware open source policies, the base image of the Dockerfile
is Photon OS, and the source code for the extension is loaded
from source instead of downloading from PyPi, for cleaner provenance.

Because the code isn't loaded by PyPi, the version
code that used pkgconfig needed to be changed.  Now
the version file will be built from environment
variables.

Finally, to build the docker images, added a makefile.
It may be a good chance to add some of the random script
files currently in the project into the makefile.

Included is a manually generated OSS tracking file.

Signed-off-by: Glenn McElhoe <gmcelhoe@vmware.com>