[VCDA-946] Updated logic to check for presence of PKS/CSE right in vCD/System org. (#267)
During CSE installation we always try to add the CSE/PKS rights and if the operation fails we look for a specific error message in the response body to infer that the right already exists in vCD. However the specific error message we are looking for is in English, which means any non English vCD installation will send us a response that will not match our criteria and the exception will leak out causing CSE installation to fail.

As a fix, we will get away from inferring the existence of the right via string matching on error messages, rather we will invoke vCD REST api to look for it directly. If the result comes back positive, we don't register the new rights, else we register them.

Also, it can be possible that the right is present in vCD but not granted to System org. Added logic to cover those cases too.

Signed-off-by: rocknes <aritra.iitkgp@gmail.com>