Change to use python files directly
Instead of downloading the container-service-extension
from PyPi in the docker image, use the python files
directly, for cleaner provenance.

Because the code isn't loaded by PyPi, the version
code that used pkgconfig needed to be changed.  Now
the version file will be built from environment
variables.

Finally, to build the docker images, added a makefile.
It may be a good chance to add some of the random script
files currently in the project into the makefile.

Signed-off-by: Glenn McElhoe <gmcelhoe@vmware.com>