VCDA-968 - 'cse run' fails when 'pks_config' is set to an empty 'pks.… (#286)
Fix to ensure that CSE server run does not crash abruptly when config files are empty. Handles exception from config checks and fails gracefully with a message.

Manually tested to ensure that CSE server gracefully fails to start when empty config files are provided to the 'cse run ' command with -c option.