Avail plans to only sys admins and removing the org vdc in use to be passed as a param for vdc to resize clusters. (#332)
This PR fixes the pks plans command option added to vcd cse ovdc list to be only available to system administrators. Also removes the org vdc in use from profiles to be passed as a param for vdc, if user does not specify a vdc as an option explicitly.

Tested vcd cse ovdc list -p as sys admin, org admin and vappauthor. Plan information is available only to system administrators.