VCDA-999: Unable to get the correct cluster config as sys admin (#361)
* Fix for VCDA-999 : Unable to get the correct cluster config as sys admin

Fixes the filtering options for system administrator and org admins for getting the cluster config correctly using the --vdc and --org options in cluster config command.

Tested both namesake clusters of type vcd and pks with the filtering options and verified that the correct config for the cluster gets retrieved. Please refer to the JIRA task for details on the bug : https://jira.eng.vmware.com/browse/VCDA-999
Additionally ran the system tests.