Change how passwords are generated and read for K8 cluster vms (#367)
While adding a node to a native K8 cluster, CSE assumes that
1. the node vm's admin password is always set.
2. the password is same as the one in the source template.

Both these assumptions can be wrong. The correct behavior in this case should be to auto generate a password while deploying the vm from template, read the password from the deployed vm instead of the template and update it to the user provided password.