[VCDA-1120] Making templates complaint with the compute policy (#410)
Templates have metadata that is derived from remote cookbook recipe and represents the template's default state. However the state can be modified during CSE server startup via template rules. This PR makes sure that once the changes are made via rules, the template is made compliant to those changed values. This PR specifically makes sure that the compute policy that is specified in the template run-time definition is actually applied to the vms in the template.