[VCDA-1199] Add K8s_version field to output of /api/cse/clusters and /api/cse/cluster/{name} (#434)
When we query vCD for listing clusters, vCD sends back metadata on those vApps too as part of the result. One of the metadata viz. template_name can be parsed to deduce the k8s version that's installed in the master node. This PR parses the name of the template from which the master node was created and appends it to the result of the call to list all clusters. If for any reason the name parsing fails, the value "Unknown" is returned.

This change only affects native clusters! To retrieve Enterprise PKS cluster's K8s version we need a separate REST call. Currently for Enterprise PKS clusters, the output of the REST call /api/clusters will have the k8s_version field set to None , while the field will be completely missing for /api/cluster/{cluster name} calls