[VCDA-1164] Introduced minor error codes in CSE responses, customization of error message on client (#436)
CSE Server side exceptions are converted into plain text error messages and sent over wire to the client. The client is normally not aware of the contents of such an error message, and can't efficiently process it to provide a more meaningful error message. To address this concern, CSE server will send a minor error code along with the error text (wherever possible), which will allow the client to identify the error accurately and make accordingly adjustments to the error message that will be displayed to end user.

Additional Fixes:
* CSE server now accepts only application/json (or any valid variant of it eg. application/*+json) requests, any other header message will result in a 406.
* Exception hierarchy has been cleaned up to better reflect the inheritance tree between related error.
* Fixed a bug in the method to add a compute policy to an ovdc which prevented a caller from any new compute policy to an ovdc.