[VCDA-1210] Fix cluster deployment for non admin cross org users with api v33.0 (#442)
In vcd api v33.0, due to a change in the way how non admin users find href of a catalog that is shared from another org, they are unable to access such catalogs and deploy k8s clusters. To fix this issue, sys admin gets the href of the catalog for these affected users.

Additional fix:
Trap exceptions raised by pyvmomi and convert them to CSE exception to patch the workflow where cluster rollback on error was being skipped.