[VCDA-1313], [VCDA-1314] Implement cluster upgrade (#495)
- Move template utility functions to local_template_manager.py
- Implement cluster upgrade client-side
- Add semantic-version dependency
- Implement cluster upgrade REST & server-side validation
- Implement cluster upgrade functionality server-side
- Move general cluster functionality outside of VcdBroker class

NOTE: cluster upgrades that only involve the Kubernetes patch version (such as increases in template revision) allow for high availability. However, to achieve this, an app must have more than 1 replica and the replicas must be running on different nodes. Facilitating this is outside of CSE's scope and must be done by the cluster admin/user

More info: https://itnext.io/keep-you-kubernetes-cluster-balanced-the-secret-to-high-availability-17edf60d9cb7