Fix indefinite wait on Guest Customization during convert-cluster operation (#506)
The REST endpoint on vCD (/vapp/vm-id/guestcustomizationstatus) that indicates whether Guest Customization has finished on a vm or not, is broken. The status never changes to anything other than GC_PENDING. As a result of this broken monitor, CSE keeps perpetually waiting on the status to change and the convert-cluster operation never finishes.

As a workaround, in this PR we have changed the waiting mechanism. Instead of waiting on the REST endpoint, we straight up look in the guest customization section of the VM, to determine if the vm representation has changed or not. This still doesn't guarantee that Guest Customization is actually complete on the vm, but it is a step in the right direction. With this change convert cluster operation will end as soon as the VM model is updated in vCD. We can still get errors like the following if we try to access the cluster right after performing convert-cluster, but waiting a minute or so will cause the issue to go away.

(cse26_beta) C:\code\cse26_beta>vcd cse cluster config test200-1
Usage: vcd cse cluster config [OPTIONS] NAME
Try "vcd cse cluster config -h" for help.

Error: (vim.fault.GuestOperationsUnavailable) {
dynamicType = ,
dynamicProperty = (vmodl.DynamicProperty) [],
msg = 'The guest operations agent could not be contacted.',
faultCause = ,
faultMessage = (vmodl.LocalizableMessage) []
}

Also changed the logic to update vm admin password to be more robust.