Revert GC fix, since fix is now available from vCD (#629)
GC fix has been submitted in vCD, hence reverting the temporary fix that was checked in via #618. This also means that the regression for photon template with older vCD builds will be fixed with this PR. However photon templates with current vCD (10.2) are still broken due to another bug in vCD. So they will remain defunct for the time being.