[VCDA-1695] Hide references to TKG PLUS unless opt-ed in (#693)
TKG PLUS is not available to CSE users by default. As a result, any reference to it should not show up on console. The two major operations where TKG PLUS related messages are dumped onto console are 1. CSE install, 2. CSE upgrade. In this PR both these workflows have been modified to suppress the messages unless a particular flag is found in the config file.