[VCDA-1890]Implement CSE server CLI to create service account
CSE on CDS doesn't have access to VCD System Administrator Role; hence
it is important to find out what all roles are necessary for CSE to
operate and provide a mechanism to provider-admin, using which such role
can be created.
This commit takes care of above requirement. As part of this requirement
it provides a new optional:
cse create-service-role
CSE server-cli.

The CLI takes care of creating CSE Service Role for CSE operations.
Optionally it also offers user to create a user of this role.

Signed-off-by: Chintan Shah <schintan@vmware.com>