[VCDA-1908] Reorder CSE installation steps, to make `creation of extension` the last step. (#840)
During CSE installation, the extension is registered before any of the other installation steps are preformed (viz. catalog creation, right registration, policy registration etc.). As a result of this if the installation fails midway, we won't be able to run cse install again, and to make matter worse since the cse install-ation process didn't complete, the next run of cse upgrade can fail due to lacking infra e.g. catalog.

To fix this issue, we are moving registration of CSE extension to be the last step during CSE installation, thereby guaranteeing that upgrade will be possible only when a valid CSE installation is in place.