[VCDA-1943] Move `adding TKG rights` from 'cse create-service-role' to 'cse install' (#875)
* [VCDA-1943] Move `adding TKG rights` from 'cse create-service-role' to 'cse install'

CSE service role command creates a role in System Org.
The rights included in this role are a hard-coded list which are
necessary for CSE to operate.
The problem here was version specific.
The hard-coded list of rights include TKG specific rights as well, which
are not always present in VCD, as they were introduced in VCD 10.2. CSE
3.0.X releases are supposed to work with VCD 10.1 as well, which do not
  have TKG rights. This creates a problem in CSE service role creation.

As a fix for this problem, we are moving this process (getting this
                rights added into CSE service role) from role creation
to CSE install operation. During CSE install operation, CSE will first
check if TKG rights are present in VCD or not, and if they are, we will
add them to service role.

Current commit takes care of this process.

Signed-off-by: Chintan Shah <schintan@vmware.com>

* [VCDA-1943] Move `adding TKG rights` from 'cse create-service-role' to 'cse install'

CSE service role command creates a role in System Org.
The rights included in this role are a hard-coded list which are
necessary for CSE to operate.
The problem here was version specific.
The hard-coded list of rights include TKG specific rights as well, which
are not always present in VCD, as they were introduced in VCD 10.2. CSE
3.0.X releases are supposed to work with VCD 10.1 as well, which do not
  have TKG rights. This creates a problem in CSE service role creation.

As a fix for this problem, we are moving this process (getting this
                rights added into CSE service role) from role creation
to CSE install operation. During CSE install operation, CSE will first
check if TKG rights are present in VCD or not, and if they are, we will
add them to service role.

Current commit takes care of this process.

Signed-off-by: Chintan Shah <schintan@vmware.com>

* [VCDA-1943] Move `adding TKG rights` from 'cse create-service-role' to 'cse install'

CSE service role command creates a role in System Org.
The rights included in this role are a hard-coded list which are
necessary for CSE to operate.
The problem here was version specific.
The hard-coded list of rights include TKG specific rights as well, which
are not always present in VCD, as they were introduced in VCD 10.2. CSE
3.0.X releases are supposed to work with VCD 10.1 as well, which do not
  have TKG rights. This creates a problem in CSE service role creation.

As a fix for this problem, we are moving this process (getting this
                rights added into CSE service role) from role creation
to CSE install operation. During CSE install operation, CSE will first
check if TKG rights are present in VCD or not, and if they are, we will
add them to service role.

Current commit takes care of this process.

Signed-off-by: Chintan Shah <schintan@vmware.com>