[VCDA-2073] Update handlers to work with request dispatcher (#935)
We are in processing of replacing request_processor.py with request_dispatcher.py. In the new module a lot of the url matching and request dispatching is state machine driven. As a result of this change, the handlers need to adapt to the new style of invocation. This PR modifies the handlers, specially the non defined entity based ones, to adhere to the new way of handling requests