[VCDA-2063, 2072] Auto negotiate api version to use from CSE client based on output of /api/cse/system (#955)
Currently CSE client blindly picks up the api version specified by the Server to communicate. However with CSE 3.1, we are expanding the number of api versions that the CSE server can concurrently support. As a result of this change the client should chose to use the highest version of api that both the server and the client understands. This PR makes the above mentioned change.

Changes:
* Updated output of /api/cse/system to contain all api versions supported by server, boolean flag to indicate if server is running in legacy mode.
* Streamlined the client code to make CSE server REST calls via the CSE_CLIENT module.
* Added logging around REST calls being made to CSE server
* Added logging around profile management by CSE client to keep track of cli command hiding
* Added logic in CSE client to negotiate the api version to use to talk to CSE server to be the max supported by the server and the client. In case the server is running in legacy mode, the api version is capped at 34.0.
* Remove 'api_version' key from config file. For the interim generate the key value pair dynamically, until all it's references are removed from server side.