[VCDA-2274] Fix upgrade issue related to missing rights in `cse service account` (#965)
CSE service account is missing the 'Edit' right for RightsBundles, as a result, it is unable to publish rights bundle to tenant orgs. This workflow will break any upgrade from CSE 2.6 or prior to CSE 3.0/CSE 3.1.

Additionally there is a logic bug around how CSE publishes the right bundle to tenants. In the current implementation, CSE doesn't preserve the list of tenants who already has the right bundle published to them. This PR fixes the bug.

Additional small fixes to
* improve flake8 compliance
* change the flag for cse create-service-role to follow standard key to ignore ssl certificate verification