Fix cse template list (#1006)
Command throws error if api_version key is not found in the config file. As of CSE 3.1 the key is not supported in the config file, so we must not be relying on it's presence. This PR removes the hard dependency, and uses the default value of None while creating pyvcloud client, which will choose the highest supported api version.

This problem only occurs if the config file is not validated before use. In case of validated config file, internally CSE sets the correct value of 'api_version' key in the run time config.