[VCDA-2198] Selective user access (#1028)
Marked Status as the protected section
Created a placeholder for the private section.
Tested the protected section thoroughly and the latest VCD build must have the fixes for all the bugs filed.

Definition of the access restrictions:

"public" -- The field is RW to all users.
"protected" -- The field is RW only to users with an FC ACL to the Entity Type. It is RO to other users.
"private" -- The field is RW only to users with an FC ACL to the Entity Type. It is not visible to other users.

Signed-off-by: sayloo <sahithi.ayloo@gmail.com>