[VCDA-2699] Split CSE MQTT extension registration and description update into two separate steps (#1118)
CSE extension is stamped with CSE specific data to record which version of CSE is being installed, which api version is CSE using etc. This update helps with future CSE upgrades.

The data stamping is generally the last step of CSE upgrade to preserve idempotency of the upgrade process itself.
This also means that if we are upgrading from a CSE system that uses AMQP to one that uses MQTT as the message bus, the MQTT exchange won't be ready until the entire upgrade process is complete.

This poses an issue since the the cluster conversions (creation/update of RDE from cluster data on vapp metadata) will not go through since in VCD 10.3, CSE native RDE operations expect the CSE MQTT exchange to be present in the system.

To solve this problem, we are splitting the AMQP -> MQTT conversion and stamping of CSE upgrade data on the extension as two separate steps. The AMQP -> MQTT conversion will take place as the first step of upgrade, while the data stamping will be the last.