[VCDA-2648] Fix NFS node addition to TKG cluster (#1132)
Users are unable to add NFS node to TKG clusters based on Ubuntu 20.04. The reason behind the failure is due to how CSE executes the nfsd.sh file on the newly created VM. Typically by the time CSE tries to run the script, the VM guest customization is over and the root password is set properly. However in case of Ubuntu 20.04, guest customization is taking way longer than usual and as a result the root password is not set properly by the time CSE tries to run the script. This leads to a failure in the vsphere-guest-run library and CSE can't recover from it.

Addition of worker nodes don't face the same issue, because CSE executes a script on master node to fetch its IP, before it tries to run any script on the worker VM. This gives the guest customization process on the worker VM enough time to run to completion.

As a fix, we are adding some extra sleep in the NFS node creation workflow to allow the VM to finish guest customization properly before trying to run nfsd.sh on it.