[VCDA 2849] Make cluster processing during CSE upgrade idempotent (#1207)
During CSE upgrade, clusters are processed and their RDE is created or updated depending on what type of cluster they are viz. legacy or RDE 1.0.0 based. Currently in CSE 3.1.0 the check to determine the cluster type is done at a high level by looking at the registered RDE (or their absence). This logic is vulnerable to changes in RDE registration (done by CSE upgrade/install) and this can lead to situations where clusters can be misclassified as RDE 1.0.0 based while they are legacy style clusters.

To fix this issue, we have changed the logic to be at cluster level. Each cluster is processed individually and the first level of check is to find the corresponding RDE, if RDE is missing it is assumed that the cluster is legacy style. If the RDE is found, it's version is determined and the corresponding update to current RDE is figured out.

Signed-off-by: rocknes <aritra.iitkgp@gmail.com>