[VCDA-3485] Update TKGm template metadata if it is renamed in VCD (#1331)
If the underlying catalog item of a TKGm template is renamed in VCD, the metadata on the catalog item goes out of sync, specifically the field name. This leads to the issue where CSE is unable to locate the affected template during cluster creation. To fix this issue, CSE will detect such name changes and automatically update the metadata. The operation will be done only for TKGm templates and will be triggered during CSE startup and whenever the template definition held in CSE server's memory is refreshed via REST api.

Signed-off-by: rocknes <aritra.iitkgp@gmail.com>